Moved event features component to a proper place
